// interfaces/statistik-spieler.ts
export interface StatistikSpieler {
    HolzMeisterMax: number;
    HolzMeisterMin: number;
    HolzMeisterSumme: number;
    HolzMeisterAVG: number;

    HolzBlitzMax: number;
    HolzBlitzMin: number;
    HolzBlitzSumme: number;
    HolzBlitzAVG: number;

    HolzMax: number;
    HolzMin: number;
    HolzSumme: number;
    HolzAVG: number;

    RattenMax: number;
    RattenSumme: number;
    NeunerMax: number;
    NeunerSumme: number;
    Kranz8Max: number;
    Kranz8Summe: number;
}